+++ CHANGELOG.md
@@ -6,6 +6,9 @@
     - Check for Instance status via the bind address, not the external one.
     - Added more status and debugging messages
     - Fixed RPC connection leak when in Grid configuration.
+    - `Node`
+        - Don't raise error if the initial neighbour is unreachable, just add
+            it to the dead list as usual.
 - `Browser`
     - Fixed issue causing the removal of cookie HttpOnly flags.
 - `Parser`